frontend/api/requirements.txt
pdfplumber==0.10.4
python-docx==1.1.0
